got the delete route working

+++ public/js/deletePost.js
@@ -1,22 +1,31 @@
-/////////////////////////////////delete a post////////////////////////////////////////////////////
+///////////////////////////////delete a post////////////////////////////////////////////////////
 const deletePost = async (event) => {
     event.preventDefault();
+    console.log('this button is working')
 
-    //get post_id from url 
-    const id = window.location.toString().split('/').at(-1)
+    if (event.target.hasAttribute('delete-id')) {
+        const id = event.target.getAttribute('delete-id')
+        console.log(id)
 
-    const response = await fetch (`/api/posts/${id}`, {
-      method: 'DELETE',
-      body:  JSON.stringify({post_id:id}),
-      headers: {'Content-Type': 'application/json'}
-    })
+        const response = await fetch (`/api/posts/${id}`, {
+            method: 'DELETE',
+          //   body:  JSON.stringify({post_id:id}),
+          //   headers: {'Content-Type': 'application/json'}
+          })
 
-    if (response.ok) {
-        document.location.reload();
-    } else {
-        alert(response.statusText)
+          if (response.ok) {
+            document.location.reload();
+        } else {
+            alert(response.statusText)
+        }
     }
+
+  
 }
+
+// function deletePost() {
+//     console.log('this button is working')
+// }
 
 document.querySelector('.delete-post-btn').addEventListener('click', deletePost)
 /////////////////////////////////////////////////////////////////////////////////////
